Switch to inheritance
